src/ui/requirements.txt
tensorflow
pillow
st-star-rating
pandas